requirements.txt
aiofiles==0.6.0
aiohttp==3.7.2
aiohttp-swagger==1.0.15
ansible==3.2.0
ansible-base==2.10.7
arrow==1.0.3
async-lru==1.0.2
async-timeout==3.0.1
attrs==20.3.0
backports.ssl==0.0.9
binaryornot==0.4.4
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
cryptography==3.3.1
dill==0.3.3
distro==1.5.0
genie==21.3.1
genie.libs.clean==21.3
genie.libs.conf==21.3
genie.libs.filetransferutils==21.3
genie.libs.health==21.3
genie.libs.ops==21.3
genie.libs.parser==21.3
genie.libs.robot==21.3
genie.libs.sdk==21.3
genie.telemetry==21.3
genie.trafficgen==21.3
idna==2.10
idna-ssl==1.1.0
importlib-metadata==3.10.0
importlib-resources==5.1.2
IxNetwork==9.0.1915.16
Jinja2==2.11.3
jinja2-time==0.2.0
jsonpickle==2.0.0
junit-xml==1.9
MarkupSafe==1.1.1
multidict==5.1.0
netaddr==0.8.0
packaging==20.9
pathspec==0.8.1
ply==3.11
poyo==0.5.0
prettytable==2.1.0
psutil==5.8.0
pyasn1==0.4.8
pyats==21.3
pyats.aereport==21.3
pyats.aetest==21.3
pyats.async==21.3
pyats.connections==21.3
pyats.contrib==21.3
pyats.datastructures==21.3
pyats.easypy==21.3
pyats.kleenex==21.3
pyats.log==21.3
pyats.reporter==21.3
pyats.results==21.3
pyats.robot==21.3
pyats.tcl==21.3
pyats.topology==21.3
pyats.utils==21.3
pycparser==2.20
pycryptodomex==3.10.1
pyftpdlib==1.5.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.1
python-engineio==3.13.2
python-slugify==4.0.1
python-socketio==4.6.0
PyYAML==5.4.1
requests==2.25.1
robotframework==4.0.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
tabulate==0.8.9
text-unidecode==1.3
tftpy==0.8.0
tqdm==4.60.0
typing-extensions==3.7.4.3
unicon==21.3
unicon.plugins==21.3
urllib3==1.26.4
virl2-client==2.1.0
wcwidth==0.2.5
websocket-client==0.58.0
xlrd==1.2.0
XlsxWriter==1.3.8
xlwt==1.3.0
xmltodict==0.12.0
yamllint==1.26.1
yarl==1.6.3
zipp==3.4.1
